Extract bundle impls from bundle.rs
